toi uu kich ban 1

+++ backend/DoAnBenhVien/DoAnBenhVien/src/main/java/com/doanbenhvien/DoAnBenhVien/DTO/DanhMucDieuTriDTO.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class DanhMucDieuTriDTO {
-    private byte idDanhMucDieuTri;
+    private Integer idDanhMucDieuTri;
     private String tenDanhMuc;
 }
